# api/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Smart-ChatBox API",
    "description": "API para gerenciamento de WhatsApp, empresas, projetos e leads.",
    "version": "1.0.0",
    "contact": {
      "name": "Tiago Marins",
      "url": "https://github.com/tiagomarins-dev/smart-chatbot-mb"
    }
  },
  "servers": [
    {
      "url": "/api",
      "description": "Servidor de produção"
    },
    {
      "url": "http://localhost/api",
      "description": "Servidor local"
    }
  ],
...
  "paths": {
...
    "/v1/api-keys": {
      "get": {
        "tags": ["api-keys"],
        "summary": "Listar chaves de API",
        "description": "Retorna a lista de chaves de API do usuário autenticado",
        "operationId": "getApiKeys",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de chaves de API",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "api_keys": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/ApiKey"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Não autorizado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["api-keys"],
        "summary": "Criar chave de API",
        "description": "Cria uma nova chave de API para o usuário autenticado",
        "operationId": "createApiKey",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "description": "Dados da chave de API",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApiKeyInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Chave de API criada com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiKeyCreationResponse"
                }
              }
            }
          },
          "400": {
            "description": "Dados inválidos",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Não autorizado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/api-keys/{id}": {
      "put": {
        "tags": ["api-keys"],
        "summary": "Atualizar chave de API",
        "description": "Atualiza uma chave de API existente",
        "operationId": "updateApiKey",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID da chave de API",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "description": "Dados de atualização da chave de API",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApiKeyUpdateInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Chave de API atualizada com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "api_key": {
                      "$ref": "#/components/schemas/ApiKey"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Dados inválidos",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Não autorizado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Chave de API não encontrada",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["api-keys"],
        "summary": "Excluir chave de API",
        "description": "Exclui permanentemente uma chave de API",
        "operationId": "deleteApiKey",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID da chave de API",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Chave de API excluída com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "API key excluída com sucesso"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Não autorizado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Chave de API não encontrada",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/api-keys/{id}/revoke": {
      "put": {
        "tags": ["api-keys"],
        "summary": "Revogar chave de API",
        "description": "Revoga (desativa) uma chave de API existente",
        "operationId": "revokeApiKey",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID da chave de API",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Chave de API revogada com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "API key revogada com sucesso"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Não autorizado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Chave de API não encontrada",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Token JWT obtido através do endpoint de autenticação"
      },
      "apiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "Chave API obtida na página de integrações"
      }
    },
    "schemas": {
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "example": "Mensagem de erro"
          }
        }
      },
      "ApiKey": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "example": "123e4567-e89b-12d3-a456-426614174000"
          },
          "user_id": {
            "type": "string",
            "format": "uuid",
            "example": "123e4567-e89b-12d3-a456-426614174000"
          },
          "name": {
            "type": "string",
            "example": "Integração Website"
          },
          "key_value": {
            "type": "string",
            "example": "api_xxxxxxxxxxxxxxxx"
          },
          "permissions": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "example": ["read:projects", "write:leads"]
          },
          "rate_limit": {
            "type": "integer",
            "example": 100
          },
          "is_active": {
            "type": "boolean",
            "example": true
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "example": "2025-05-04T12:00:00Z"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "example": "2025-05-04T12:00:00Z"
          },
          "expires_at": {
            "type": "string",
            "format": "date-time",
            "example": "2026-05-04T12:00:00Z"
          },
          "last_used_at": {
            "type": "string",
            "format": "date-time",
            "example": "2025-05-06T15:30:00Z"
          }
        }
      },
      "ApiKeyInput": {
        "type": "object",
        "required": ["name"],
        "properties": {
          "name": {
            "type": "string",
            "example": "Integração Website"
          },
          "permissions": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "example": ["read:projects", "write:leads"]
          },
          "rate_limit": {
            "type": "integer",
            "example": 100
          },
          "expires_at": {
            "type": "string",
            "format": "date-time",
            "example": "2026-05-04T12:00:00Z"
          }
        }
      },
      "ApiKeyCreationResponse": {
        "type": "object",
        "properties": {
          "api_key": {
            "$ref": "#/components/schemas/ApiKey"
          },
          "secret": {
            "type": "string",
            "description": "Secret da API key (exibido apenas uma vez)",
            "example": "secret_xxxxxxxxxxxxxxxxxxxxxxxx"
          },
          "warning": {
            "type": "string",
            "example": "O segredo (secret) não será exibido novamente. Salve-o em um local seguro."
          }
        }
      },
      "ApiKeyUpdateInput": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Integração Website Atualizada"
          },
          "permissions": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "example": ["read:projects", "write:leads", "read:companies"]
          },
          "rate_limit": {
            "type": "integer",
            "example": 200
          },
          "expires_at": {
            "type": "string",
            "format": "date-time",
            "example": "2026-05-04T12:00:00Z"
          },
          "is_active": {
            "type": "boolean",
            "example": true
          }
        }
      }
    }
  }
}